refactor: js 파일 역할에 따라 파일 분리
