+++ README.md
@@ -68,7 +68,8 @@
 - [x] 구매금액 입력 후 enter 시, 클릭 효과 적용
 - [ ] html 구조 리팩토링 (클래스 및 id 명 통일, 시맨틱 태그 적용)
   - [x] html 파일 수정
-  - [ ] js 동적 생성 (당첨 통계 모달 내부 내용)
+- [x] js 역할에 따라 파일 분리
+  - [ ] 공통 로직 -> utils로 분리 (초기화 하는 코드 )
 - [ ] dom 요소 select 파일 분리
 - [ ] `MATCH_COUNT` constant 분리 (갯수, 가격 - 중첩 객체)
 - [ ] 인스턴스 변수 줄이기 - (`LottoPack` 클래스 현재 3개)
